feat: se standarizan los nombres y se agrega la cotizacion de banco basa

+++ src/main/java/py/com/cotip/external/cotipdb/entities/CotipEntity.java
@@ -21,9 +21,9 @@ public class CotipEntity extends CotipBaseEntity {
 
     private String currencyCode;
 
-    private BigDecimal buyRate;
+    private long buyRate;
 
-    private BigDecimal sellRate;
+    private long sellRate;
 
 
 }
